+++ src/index.ts
@@ -1,4 +1,13 @@
 export * from './k-number';
 export * from './constants';
 export * from './errors';
-export * from './types';
+export * from './utils';
+export type {
+  NumberString,
+  NonZeroNumberString,
+  KoreanNumberString,
+  KoreanUnit,
+  KNumberFormat,
+  NumberAndKoreanRecord,
+  KNumberConfig,
+} from './types';